Crate canyon_inspect

A CLI extension for substrate node, adding sub-command to pretty print debug info about blocks and extrinsics.

The blocks and extrinsics can either be retrieved from the database (on-chain), or a raw SCALE-encoding can be provided.

Modules

cli
Structs to easily compose inspect sub-command for CLI.

command
Command ran by the CLI

Structs

DebugPrinter
Default dummy debug printer.

Inspector
Blockchain inspector.

Enums

BlockAddress
A block to retrieve.

Error
Aggregated error for Inspector operations.

ExtrinsicAddress
An extrinsic address to decode and print out.

Traits

ChainAccess
A helper trait to access block headers and bodies.

PrettyPrinter
A Pretty formatter implementation.

Type Definitions

BlockAddressFor
A helper type for a generic block input.